Squashed commit of the following:
commit acc04091160ae2a82d295feac0ff7c8681e7486c
Merge: 677b823 0ef9fa6
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sun May 20 11:23:34 2018 +0200

    Merge branch 'master' into BR-services

commit 677b823ab4b412527ca14318f31056ce2d3f2449
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sun May 20 11:14:44 2018 +0200

    Dry option

commit bbea5e018d7e0030b8c3e9de2f2674544566e484
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sun May 20 10:47:32 2018 +0200

    Cleanup

commit fb9deab8309abb6629caf1db74d6983647b653bf
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 22:49:19 2018 +0200

    Dependency update

commit cdc5ef6f8e4ceadd4eb0ec75c8b70fc0d6729487
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 22:48:34 2018 +0200

    Reformat

commit ab645080fd0ce2c4259c00342ecdba49312f780a
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 22:44:30 2018 +0200

    Backup commands, exceptions, debug info, reformat

commit 89e8d4f146c3388b3c6b1c621f6231ef00cc3f8a
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 22:34:17 2018 +0200

    Add consoles

commit ea57248a3589a2ca57a2d548517eae4ca2374d78
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 22:33:21 2018 +0200

    Storage and backups

commit fa07c4b6493d96db2fa92a3f3de851421459a201
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 18:55:14 2018 +0200

    Remove comments

commit 641b2b467883e58ce5fcefde0c848f32866269c5
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 18:26:17 2018 +0200

    Dependency update

commit 65b6091cdbf63d3d4d7e6672e99ca220a11c0445
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 18:24:55 2018 +0200

    Migrations

commit bc0bf06b12b9229103caaebea3ded685d269f9d9
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 18:09:18 2018 +0200

    Dump only if active

commit b5a3e962218ac4088ef3eca2b756610dcb82ee81
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 17:58:56 2018 +0200

    Empty services by default

commit 2cbb96294a46d40cbc8ed56773f0a3d6aedbfb92
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 13:37:52 2018 +0200

    Dependency update

commit 5197bac72975550b187cbdb4454907e8a0314c2b
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 19 13:35:55 2018 +0200

    Pass server when creating or changing service for context

commit a3f119b8edd7c0236ea1b15798c9072dfafdf5e3
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Fri May 18 20:23:19 2018 +0200

    Cleanup

commit bae6b9c9a4ba878974084886498dce7393bed382
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Fri May 18 17:24:24 2018 +0200

    Changes

commit bff6b047cdb22b11d5b5571f73de7ed73bac58cc
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Fri May 18 16:45:22 2018 +0200

    Upload to spaces ;)

commit 67ecfd558a148abcaced369e7fab7b821620bfee
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Thu May 17 21:30:12 2018 +0200

    Services

commit dfd50f1a76ccc2ea7b4693c74d487bd3b90bbef1
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Thu May 17 18:50:32 2018 +0200

    Dependency update

commit 7b8bd9ad4896d00c1652455a3955772c11a2ba94
Merge: 6305d60 3090791
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Thu May 17 18:39:54 2018 +0200

    Merge branch 'master' into BR-services

commit 6305d6017364e5f4678a364d9d967e915969a56e
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Thu May 17 12:01:52 2018 +0200

    Cronjobs

commit 153e9dac9490c2d64dbb01cc63cd98dbaa983b36
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 16 14:35:18 2018 +0200

    Docblock

commit bd313db9dc7a8d1d2c30ea163c84a679918c55ce
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 16 14:22:45 2018 +0200

    Some comments

commit 975243c05064d111c1a0801a5f7dc6bddb73a237
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 16 14:20:44 2018 +0200

    Remove things a little bit

commit c866024e122ba5d5a32675c1d2e7fc5ebc220b2d
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 16 14:15:49 2018 +0200

    Slowly coming to a point :)

commit 5d98a199a846d91095ddde25d1b2c2989a164bd9
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 16 13:41:05 2018 +0200

    spaces cold storage

commit d4af2bd0a5ad50383e787b307b351a2845be264d
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 16 12:09:37 2018 +0200

    Refactor crypto

commit 9b3f5253c60dd73ea1701e79d583924a031ed404
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Tue May 15 15:51:47 2018 +0200

    GPG, openssl, backup, export, import, ...

commit 334e29d1364febfc036095f8b124b2f6937b5949
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 12 19:31:10 2018 +0200

    Renames

commit 1209e7bbd1939fd6d5747aa7a9fffb4208058618
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 12 19:21:19 2018 +0200

    Email

commit f1338a5f8ad1eb92485f379037a061d550ae8ab3
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 12 19:10:50 2018 +0200

    Dependency update

commit aaa68c4f14cc2819afd6b6b48ac44a1e61efaf08
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 12 19:10:25 2018 +0200

    Dependency update

commit acba245df90c22c4a744ec8b64909e3c32d21e49
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Sat May 12 19:00:32 2018 +0200

    Openssl encryption, tons of stuff

commit e4a55df68845503fc29405b0c81f510cd47dc4d6
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 9 20:03:59 2018 +0200

    Encryption secrets, delete platform request

commit a2c50a7b0316b4c5c662c1be4f9fdd835ed522be
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 9 15:29:33 2018 +0200

    Few changes

commit 57935df695fb53154d25aa25a0e55e799a8756c0
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 9 15:13:08 2018 +0200

    Replication and backups, stil not tested

commit 884972bea50d03bc877b80641b33d335003f77fc
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Wed May 9 13:31:18 2018 +0200

    Refactor some services

commit a0de89568e0c96c6d13745d0485653b4463cee5c
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Tue May 8 22:27:10 2018 +0200

    Database replication, backup, compression, encryption, ... alpha, not tested

commit 58ed85900059e7331cd0d826938267de8693492c
Merge: 79339c0 6f80074
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Fri May 4 14:59:43 2018 +0200

    Merge branch 'master' into BR-services

commit 79339c0dbae8019d641104369705bfcec9e27423
Author: Bojan Rajh <schtr4jh@schtr4jh.net>
Date:   Mon Apr 2 22:25:18 2018 +0200

    Services

+++ app/impero/src/Impero/Apache/Console/DumpVirtualhosts.php
@@ -1,7 +1,5 @@
 <?php namespace Impero\Apache\Console;
 
-use Impero\Apache\Entity\Sites;
-use Impero\Apache\Record\Site;
 use Impero\Servers\Entity\Servers;
 use Impero\Servers\Record\Server;
 use Pckg\Framework\Console\Command;
@@ -10,6 +8,9 @@ use Symfony\Component\Console\Input\InputOption;
 class DumpVirtualhosts extends Command
 {
 
+    /**
+     * @throws \Symfony\Component\Console\Exception\InvalidArgumentException
+     */
     protected function configure()
     {
         $this->setName('apache:dump')
@@ -19,9 +20,18 @@ class DumpVirtualhosts extends Command
                      'server' => 'Server ID',
                  ],
                  InputOption::VALUE_REQUIRED
+             )
+             ->addOptions(
+                 [
+                     'dry' => 'Do not dump or restart',
+                 ],
+                 InputOption::VALUE_NONE
              );
     }
 
+    /**
+     * @throws \Exception
+     */
     public function handle()
     {
         if (!$this->option('server')) {
@@ -30,45 +40,49 @@ class DumpVirtualhosts extends Command
             return;
         }
 
-        $this->output('Building virtualhosts');
+        /**
+         * Get server.
+         */
         $server = (new Servers())->where('id', $this->option('server'))->oneOrFail();
-        $sites = (new Sites())->where('server_id', $server->id)->all();
-        $virtualhosts = [];
-        //$virtualhostsNginx = [];
-        $sites->each(
-            function(Site $site) use ($server, &$virtualhosts/*, &$virtualhostsNginx*/) {
-                /**
-                 * Apache: apache port
-                 * Nginx: nginx port
-                 * Haproxy: haproxy port
-                 */
-                if ($site->hasServiceOnServer($server, 'apache')) {
-                    $virtualhosts[] = $site->getVirtualhost();
-                }
-                if ($site->hasServiceOnServer($server, 'nginx')) {
-                }
-                if ($site->hasServiceOnServer($server, 'haproxy')) {
-                }
-                //$virtualhostsNginx[] = $site->getVirtualhostNginx();
-            }
-        );
 
-        $virtualhosts = implode("\n\n", $virtualhosts);
-        //$virtualhostsNginx = implode("\n\n", $virtualhostsNginx);
+        /**
+         * Get server services: web and lb.
+         *
+         * @T00D00 - run this in parallel
+         */
 
-        $this->output('Dumping virtualhosts');
+        $this->output('Building apache');
+        $virtualhosts = $server->getApacheConfig();
 
+        $this->output('Building haproxy');
+        $virtualhostsHaproxy = $server->getHaproxyConfig();
+
+        $this->output('Dumping apache');
         $this->storeVirtualhosts($server, $virtualhosts);
-        //$this->storeVirtualhostsNginx($server, $virtualhostsNginx);
 
-        $this->output('Virtualhosts were dumped, waiting for apache graceful');
+        if ($server->getSettingValue('service.haproxy.active')) {
+            $this->output('Dumping haproxy');
+            $this->storeVirtualhostsHaproxy($server, $virtualhostsHaproxy);
+        }
+
+        $this->output('Done');
     }
 
+    /**
+     * @param Server $server
+     * @param        $virtualhosts
+     *
+     * @throws \Exception
+     */
     protected function storeVirtualhosts(Server $server, $virtualhosts)
     {
         $local = '/tmp/server.' . $server->id . '.virtualhosts';
         $remote = '/etc/apache2/sites-enabled/002-impero.conf';
         file_put_contents($local, $virtualhosts);
+        if ($this->option('dry')) {
+            return;
+        }
+        $this->outputDated('Dumping and restarting (apache)');
         $sshConnection = $server->getConnection();
         $sshConnection->sftpSend($local, $remote);
         unlink($local);
@@ -79,12 +93,19 @@ class DumpVirtualhosts extends Command
         $sshConnection->exec('sudo service apache2 graceful');
     }
 
+    /**
+     * @param Server $server
+     * @param        $virtualhosts
+     */
     protected function storeVirtualhostsNginx(Server $server, $virtualhosts)
     {
-        return;
         $local = '/tmp/server.' . $server->id . '.virtualhosts';
         $remote = '/etc/apache2/sites-enabled/002-impero.conf';
         file_put_contents($local, $virtualhosts);
+        if ($this->option('dry')) {
+            return;
+        }
+        $this->outputDated('Dumping and restarting (nginx)');
         $sshConnection = $server->getConnection();
         $sshConnection->sftpSend($local, $remote);
         unlink($local);
@@ -93,6 +114,28 @@ class DumpVirtualhosts extends Command
          * @T00D00 - check if apache is offline and apply previous configuration.
          */
         $sshConnection->exec('sudo service nginx restart');
+    }
+
+    /**
+     * @param Server $server
+     * @param        $virtualhosts
+     *
+     * @throws \Exception
+     */
+    protected function storeVirtualhostsHaproxy(Server $server, $virtualhosts)
+    {
+        $local = '/tmp/server.' . $server->id . '.haproxy';
+        $remote = '/etc/haproxy/haproxy.cnf';
+        file_put_contents($local, $virtualhosts);
+        if ($this->option('dry')) {
+            return;
+        }
+        $this->outputDated('Dumping and restarting (haproxy)');
+        $sshConnection = $server->getConnection();
+        $sshConnection->sftpSend($local, $remote);
+        unlink($local);
+
+        $sshConnection->exec('sudo service haproxy restart');
     }
 
 }
